# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://10.114.34.65/"
    },
    {
      "url": "https://10.114.34.65/"
    }
  ],
  "paths": {
...
    "/ideas/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/idea"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/idea"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/idea"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/idea"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/idea"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/idea"
                }
              }
            }
          }
        }
      }
    },
    "/likes/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/likeFull"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/likeFull"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idea_id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/likes/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              }
            }
          }
        }
      }
    },
    "/likes/idea/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/idea"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/idea"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/likeFull"
                }
              }
            }
          }
        }
      }
    },
...
    "/comments/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "idea": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Add coffee machine"
          },
          "content": {
            "type": "string",
            "example": "We really should have access to free coffee."
          },
          "created_at": {
            "type": "string",
            "example": "2022-11-23T17:52:40.243Z"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 10
              },
              "name": {
                "type": "string",
                "example": "Victor Mike"
              }
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "Nice idea"
                },
                "id": {
                  "type": "number",
                  "example": 2
                },
                "updated_at": {
                  "type": "string",
                  "example": "2022-11-24T10:36:10.190Z"
                },
                "idea": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    }
                  }
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 10
                    },
                    "name": {
                      "type": "string",
                      "example": "Victor Mike"
                    }
                  }
                }
              }
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "idea_id": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "tag": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Cafeteria"
                    },
                    "id": {
                      "type": "number",
                      "example": 15
                    },
                    "description": {
                      "type": "string",
                      "example": "Ideas related to food"
                    }
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "idea"
        }
      },
...
      "comment": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Nice idea"
          },
          "id": {
            "type": "number",
            "example": 2
          },
          "updated_at": {
            "type": "string",
            "example": "2022-11-24T10:36:10.190Z"
          },
          "idea": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              }
            }
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 10
              },
              "name": {
                "type": "string",
                "example": "Victor Mike"
              }
            }
          }
        },
        "xml": {
          "name": "comment"
        }
      },
...
      "likeFull": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 10
              },
              "name": {
                "type": "string",
                "example": "Victor Mike"
              }
            }
          },
          "id": {
            "type": "number",
            "example": 30
          },
          "idea": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              },
              "user_id": {
                "type": "number",
                "example": 1
              }
            }
          },
          "created_at": {
            "type": "string",
            "example": "2022-11-30T19:01:28.126Z"
          }
        },
        "xml": {
          "name": "likeFull"
        }
      },
...
    }
  }
}